Split carver operator per-tool
